Implementación y cambios en mooringCategoryService

+++ src/main/java/com/nautik/api/repository/port/PortRepository.java
@@ -16,13 +16,13 @@ public interface PortRepository extends JpaRepository<Port, Integer> {
     Optional<Port> findByNameIgnoreCase(String name);
 
     @Query("select max(d.maxLength) from MooringDimension d inner join Port p on d.port = p where p.id = ?1")
-    Integer getMaxLengthByPortId(Integer portId);
+    Double getMaxLengthByPortId(Integer portId);
 
     @Query("select max(d.maxBeam) from MooringDimension d inner join Port p on d.port = p where p.id = ?1")
-    Integer getMaxBeamByPortId(Integer portId);
+    Double getMaxBeamByPortId(Integer portId);
 
     @Query("select max(d.maxDraft) from MooringDimension d inner join Port p on d.port = p where p.id = ?1")
-    Integer getMaxDraftByPortId(Integer portId);
+    Double getMaxDraftByPortId(Integer portId);
 
     @Query("select count(m) from Mooring m where m.mooringCategory.zone.port.id = ?1")
     Integer mooringNumberOfPort(Integer portId);
